Update code_of_conduct.md
## Statement of Intent

The wording of " individual teams" is a little bit ironic. Edited out that second sentence.

Also created a few changes to that section to make it flow better. I think the general message is the same, just more concise. Feel free to change though if there's anything else you think should be added/changed :)

## Code of Conduct

Changed the last sentence to:
- "By the nature of our decentralized space, the adoption of this code of conduct will be voluntary. It is up to Aragon Cooperative members to identify, discuss, and enforce consequences when inappropriate conduct occurs. "

## Feedback 

We don't have a documentation working group. I don't think we're ready for one yet. Removed that sentence.

It makes sense to have a place to keep this stuff. I think the simplest way to do that is to create a `docs` folder in the temp repo, then move all this stuff in there. That way the current Meta Proposal process can be used to amend the current CoC or whatever else. 

I think it would be good to have a place/way for people to discretely provide feedback on inappropriate conduct and get it resolved. The Ethereum Magicians had lots of thoughts on this, but I don't know if any materialized. For our purposes, maybe it makes sense to have a few designated "community managers" who can help resolve disputes and address inappropriate conduct? 

<br>

Otherwise everything looks good for an initial CoC doc.